Add test for BasicAsset edit operation
Test BasicAsset edit operation

This PR adds a unit test to verify the `_edit_asset` method of the `BasicAsset` class.

Changes:
- Implement `test_edit_asset` method
- Mock the `asset_update` API call
- Validate that the `_edit_asset` method correctly invokes the API with the provided asset

The test ensures that the `_edit_asset` method properly interacts with the underlying asset update API to save changes to the asset. This is an important part of the asset management functionality.